Checking in just a merge of one revision in the to-release-as-1.8 branch: ------------------------------------------------------------------------ r3687 | billvenners@gmail.com | 2012-04-03 20:20:30 -0700 (Tue, 03 Apr 2012) | 3 lines
Merged in Chee Seng's 18-cs2 branch, which added an overloaded apply method to Informer that takes a
payload.

The command I used was:

svn merge -r3686:3687 https://scalatest.googlecode.com/svn/branches/to-release-as-1.8

In this one, there is an overloaded apply method that takes an Any payload. Later I change that
to a single apply method with a Option[Any] payload with a default of None.



git-svn-id: http://scalatest.googlecode.com/svn/trunk@3749 fd1274fd-080f-5963-8033-43ee40fcd1de