Merged in the remaining changes made to the to-release-as-1.8 branch. The command I used was:
svn merge -r3733:3823 https://scalatest.googlecode.com/svn/branches/to-release-as-1.8

I had 7 conflicts, which I resolved in various ways. For build.xml and ScalaTestFramework.scala, I just reverted
back to the original trunk version. We'll need to by-hand implement the changes made to ScalaTestFramework, which
appears to be adding -y, -F, and -P support.
I used the new AppendedCluesSpec.scala and AppendedClues.scala, because those started with the trunk version.
I merged Runner.scala, ScalaTestAntTask.scala, and ConcurrentDistributor.scala, but went back to the trunk
version of ConcurrentDistributor.scala in the end (I think), because it is tied up with passing filter into
apply, and that's wrong. Will need to fix that in Runner and then fix ConcurrentDistributor (go back to the
regular apply that's inherited from Distributor.)



git-svn-id: http://scalatest.googlecode.com/svn/trunk@3854 fd1274fd-080f-5963-8033-43ee40fcd1de